refactor: 카테고리 중분류, 소분류 셀렉트박스 리펙토링
