+++ src/main/java/com/donut/prokindonutsweb/homeController/HomeController.java
@@ -35,7 +35,4 @@ public class HomeController {
     @GetMapping("/qh/outbound")
     public void Qoutbound(){}
 
-    @GetMapping("/qh/product")
-    public void Qproduct(){}
-
 }
