+++ src/main/java/com/donut/prokindonutsweb/inventory/controller/QhInventoryController.java
@@ -40,16 +40,4 @@ public class QhInventoryController {
     model.addAttribute("categoryMidList", categoryMidList);
   }
   
-  @GetMapping(value = "/check", produces = "application/json; charset=UTF-8")
-  @ResponseBody
-  public List<String> getSubCategoryList(@RequestParam("categoryMid") String categoryMid) {
-    
-    if (categoryMid == null || categoryMid.trim().isEmpty() || "중분류 선택".equals(categoryMid)) {
-      System.out.println("test : " + categoryMid);
-      return List.of();
-    }
-    
-    return categoryFilterService.findSubCategoryListByMid(categoryMid);
-  }
-  
 }
