+++ src/main/webapp/WEB-INF/views/wm/warehouse.jsp
@@ -308,7 +308,7 @@
 
             // 3-2. fetch 요청
             const contextPath = "${pageContext.request.contextPath}";
-            fetch(`${contextPath}/qh/inventory/check?categoryMid=`+encodeURIComponent(midVal))
+            fetch(`${contextPath}/category/check?categoryMid=`+encodeURIComponent(midVal))
                 .then(res => res.json())
                 .then(subList => {
                     subList.forEach(sub => {
